requirements.txt
pydantic-ai
python-dotenv
jinja2
ipykernel 